add listners and upgrade crewai deps
